# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/banner": {
            "get": {
                "description": "Получает список баннеров с учетом переданных параметров запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить баннеры с учетом заданных параметров",
                "operationId": "get-banners",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор особенности баннера",
                        "name": "feature_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор тега баннера",
                        "name": "tag_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ограничение на количество возвращаемых баннеров",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение для пагинации результатов",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список баннеров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Banner"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении баннеров",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
        "/banner/{id}": {
            "delete": {
                "description": "Удаляет баннер с указанным идентификатором",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Удалить баннер по его идентификатору",
                "operationId": "delete-banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор баннера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Баннер успешно удален"
                    },
                    "400": {
                        "description": "Неверный идентификатор баннера",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка при удалении баннера",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет информацию о баннере с указанным идентификатором",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Обновить информацию о баннере",
                "operationId": "update-banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Идентификатор баннера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления баннера",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/banners.RequestUpdateBanner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный баннер",
                        "schema": {
                            "$ref": "#/definitions/banners.ResponseBanner"
                        }
                    },
                    "400": {
                        "description": "Неверный идентификатор баннера или неверные данные запроса",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "404": {
                        "description": "Баннер не найден",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    },
                    "500": {
                        "description": "Ошибка при обновлении баннера",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
        "/banners": {
            "post": {
                "description": "Создает новый баннер на основе переданных данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Создать новый баннер",
                "operationId": "create-banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен администратора",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для создания баннера",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/banners.RequestBanner"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданный баннер",
                        "schema": {
                            "$ref": "#/definitions/banners.ResponseBanner"
                        }
                    },
                    "400": {
                        "description": "Ошибка при создании баннера",
                        "schema": {
                            "$ref": "#/definitions/banners.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "banners.RequestBanner": {
            "type": "object",
            "required": [
                "content",
                "feature_id",
                "is_active",
                "tag_ids"
            ],
            "properties": {
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "feature_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "banners.RequestUpdateBanner": {
            "type": "object",
            "required": [
                "content",
                "feature_id",
                "is_active",
                "tag_ids"
            ],
            "properties": {
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "feature_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "banners.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "banners.ResponseBanner": {
            "type": "object",
            "properties": {
                "banner_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "error": {
                    "type": "string"
                },
                "feature_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "entities.Banner": {
            "type": "object",
            "properties": {
                "banner_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "created_at": {
                    "type": "string"
                },
                "feature_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}